FIX: `writeFileSafely`で一時ファイル名にUUIDを使用し、重複を避ける

+++ src/backend/electron/fileHelper.ts
@@ -1,6 +1,7 @@
 import fs from "fs";
 import log from "electron-log/main";
 import { moveFileSync } from "move-file";
+import { uuid4 } from "@/helpers/random";
 
 /**
  * 書き込みに失敗したときにファイルが消えないように、
@@ -10,34 +11,19 @@ export function writeFileSafely(
   path: string,
   data: string | NodeJS.ArrayBufferView,
 ) {
-  let tmpPath: string;
-  let maxRetries = 16;
-  while (true) {
-    maxRetries--;
-    try {
-      // ランダムな文字列を8文字生成
-      const randStr = Math.floor(Math.random() * 36 ** 8)
-        .toString(36)
-        .padStart(8, "0");
-      tmpPath = `${path}-${randStr}.tmp`;
-      fs.writeFileSync(tmpPath, data, { flag: "wx" });
-      break;
-    } catch (error) {
-      const e = error as NodeJS.ErrnoException;
-      if (e.code !== "EEXIST" || maxRetries <= 0) {
-        throw e;
-      }
-    }
-  }
+  const tmpPath = `${path}-${uuid4()}.tmp`;
+  fs.writeFileSync(tmpPath, data);
 
   try {
     moveFileSync(tmpPath, path, {
       overwrite: true,
     });
   } catch (error) {
-    fs.promises.unlink(tmpPath).catch((reason) => {
-      log.warn("Fail to remove %s\n  %o", tmpPath, reason);
-    });
+    if (fs.existsSync(tmpPath)) {
+      fs.promises.unlink(tmpPath).catch((reason) => {
+        log.warn("Failed to remove %s\n  %o", tmpPath, reason);
+      });
+    }
     throw error;
   }
 }
